Chatbot-with-FastAPI-and-Memory/requirements.txt
# Backend
fastapi
uvicorn
pydantic
python-dotenv
groq   # Groq API SDK

# Frontend
requests
streamlit